Refactor PostgreSql client into partial files
