+++ DbaClientX.PostgreSql/PostgreSql.Transactions.cs
@@ -1,0 +1,260 @@
+using System;
+using System.Data;
+using System.Threading;
+using System.Threading.Tasks;
+using Npgsql;
+
+namespace DBAClientX;
+
+public partial class PostgreSql
+{
+    /// <summary>
+    /// Begins a new transaction using the default isolation level.
+    /// </summary>
+    public virtual void BeginTransaction(string host, string database, string username, string password)
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                throw new DbaTransactionException("Transaction already started.");
+            }
+
+            var connectionString = BuildConnectionString(host, database, username, password);
+
+            _transactionConnection = new NpgsqlConnection(connectionString);
+            _transactionConnection.Open();
+            _transaction = _transactionConnection.BeginTransaction();
+        }
+    }
+
+    /// <summary>
+    /// Begins a new transaction using the specified isolation level.
+    /// </summary>
+    public virtual void BeginTransaction(string host, string database, string username, string password, IsolationLevel isolationLevel)
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                throw new DbaTransactionException("Transaction already started.");
+            }
+
+            var connectionString = BuildConnectionString(host, database, username, password);
+
+            _transactionConnection = new NpgsqlConnection(connectionString);
+            _transactionConnection.Open();
+            _transaction = _transactionConnection.BeginTransaction(isolationLevel);
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously begins a new transaction using the default isolation level.
+    /// </summary>
+    public virtual async Task BeginTransactionAsync(string host, string database, string username, string password, CancellationToken cancellationToken = default)
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                throw new DbaTransactionException("Transaction already started.");
+            }
+        }
+
+        var connectionString = BuildConnectionString(host, database, username, password);
+
+        var connection = new NpgsqlConnection(connectionString);
+        await connection.OpenAsync(cancellationToken).ConfigureAwait(false);
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER || NET5_0_OR_GREATER
+        var transaction = await connection.BeginTransactionAsync(cancellationToken).ConfigureAwait(false);
+#else
+        var transaction = connection.BeginTransaction();
+#endif
+
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                transaction.Dispose();
+                connection.Dispose();
+                throw new DbaTransactionException("Transaction already started.");
+            }
+
+            _transactionConnection = connection;
+            _transaction = transaction;
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously begins a new transaction using the specified isolation level.
+    /// </summary>
+    public virtual async Task BeginTransactionAsync(string host, string database, string username, string password, IsolationLevel isolationLevel, CancellationToken cancellationToken = default)
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                throw new DbaTransactionException("Transaction already started.");
+            }
+        }
+
+        var connectionString = BuildConnectionString(host, database, username, password);
+
+        var connection = new NpgsqlConnection(connectionString);
+        await connection.OpenAsync(cancellationToken).ConfigureAwait(false);
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER || NET5_0_OR_GREATER
+        var transaction = await connection.BeginTransactionAsync(isolationLevel, cancellationToken).ConfigureAwait(false);
+#else
+        var transaction = connection.BeginTransaction(isolationLevel);
+#endif
+
+        lock (_syncRoot)
+        {
+            if (_transaction != null)
+            {
+                transaction.Dispose();
+                connection.Dispose();
+                throw new DbaTransactionException("Transaction already started.");
+            }
+
+            _transactionConnection = connection;
+            _transaction = transaction;
+        }
+    }
+
+    /// <summary>
+    /// Commits the currently active transaction.
+    /// </summary>
+    public virtual void Commit()
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            _transaction.Commit();
+            DisposeTransactionLocked();
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously commits the currently active transaction.
+    /// </summary>
+    public virtual async Task CommitAsync(CancellationToken cancellationToken = default)
+    {
+        NpgsqlTransaction? tx;
+        NpgsqlConnection? conn;
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            tx = _transaction;
+            conn = _transactionConnection;
+            _transaction = null;
+            _transactionConnection = null;
+        }
+
+        try
+        {
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER || NET5_0_OR_GREATER
+            await tx!.CommitAsync(cancellationToken).ConfigureAwait(false);
+#else
+            tx!.Commit();
+#endif
+        }
+        finally
+        {
+            tx!.Dispose();
+            conn?.Dispose();
+        }
+    }
+
+    /// <summary>
+    /// Rolls back the currently active transaction.
+    /// </summary>
+    public virtual void Rollback()
+    {
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            _transaction.Rollback();
+            DisposeTransactionLocked();
+        }
+    }
+
+    /// <summary>
+    /// Asynchronously rolls back the currently active transaction.
+    /// </summary>
+    public virtual async Task RollbackAsync(CancellationToken cancellationToken = default)
+    {
+        NpgsqlTransaction? tx;
+        NpgsqlConnection? conn;
+        lock (_syncRoot)
+        {
+            if (_transaction == null)
+            {
+                throw new DbaTransactionException("No active transaction.");
+            }
+
+            tx = _transaction;
+            conn = _transactionConnection;
+            _transaction = null;
+            _transactionConnection = null;
+        }
+
+        try
+        {
+#if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_0_OR_GREATER || NET5_0_OR_GREATER
+            await tx!.RollbackAsync(cancellationToken).ConfigureAwait(false);
+#else
+            tx!.Rollback();
+#endif
+        }
+        finally
+        {
+            tx!.Dispose();
+            conn?.Dispose();
+        }
+    }
+
+    private void DisposeTransaction()
+    {
+        lock (_syncRoot)
+        {
+            DisposeTransactionLocked();
+        }
+    }
+
+    private void DisposeTransactionLocked()
+    {
+        _transaction?.Dispose();
+        _transaction = null;
+        _transactionConnection?.Dispose();
+        _transactionConnection = null;
+    }
+
+    /// <inheritdoc />
+    protected override bool IsTransient(Exception ex) =>
+        ex is PostgresException pgEx &&
+        pgEx.SqlState is "40001" or "40P01" or "55P03" or "53300" or "55006";
+
+    /// <inheritdoc />
+    protected override void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            DisposeTransaction();
+        }
+
+        base.Dispose(disposing);
+    }
+}
